third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-disconnect-called-during.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => {
      return gatt.getPrimaryService('heart_rate')
        .then(heart_rate_service => {
          let promise = assert_promise_rejects_with_message(
            heart_rate_service.getCharacteristics(),
            new DOMException(
              'GATT Server is disconnected. Cannot retrieve characteristics. ' +
              '(Re)connect first with `device.gatt.connect`.',
              'NetworkError'));
          gatt.disconnect();
          return promise;
        });
    });
}, 'disconnect() called during getCharacteristics. Reject with NetworkError.');

</script>
